RepositoryInfoの切り出し
RepositoryInfoの切り出し
