+++ dashboard/src/pages/apps/new.tsx
@@ -30,6 +30,7 @@ import { createStore, SetStoreFunction } from 'solid-js/store'
 import toast from 'solid-toast'
 import { ConnectError } from '@bufbuild/connect'
 import { storify } from '/@/libs/storify'
+import { RepositoryInfo } from '/@/components/RepositoryInfo'
 
 const [repos] = createResource(() => client.getRepositories({}))
 const [apps] = createResource(() => client.getApplications({}))
@@ -373,60 +374,6 @@ const FormRuntimeConfig = (props: FormRuntimeConfigProps) => {
         />
       </Form>
     </>
-  )
-}
-
-const RepositoryInfoContainer = styled('div', {
-  base: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: '10px',
-  },
-})
-
-const RepoName = styled('div', {
-  base: {
-    fontSize: '16px',
-    fontWeight: 500,
-    color: vars.text.black1,
-  },
-})
-
-export interface RepositoryInfoProps {
-  repo: Repository
-  apps: Application[]
-}
-
-const RepositoryInfoBackground = styled('div', {
-  base: {
-    display: 'flex',
-
-    background: vars.bg.white3,
-    border: `1px solid ${vars.bg.white4}`,
-    borderRadius: '4px',
-    padding: '8px 12px',
-  },
-})
-
-const SmallText = styled('div', {
-  base: {
-    display: 'flex',
-    fontSize: '11px',
-    color: vars.text.black3,
-  },
-})
-
-const RepositoryInfo = (props: RepositoryInfoProps): JSXElement => {
-  const provider = repositoryURLToProvider(props.repo.url)
-  return (
-    <RepositoryInfoBackground>
-      <RepositoryInfoContainer>
-        {providerToIcon(provider)}
-        <RepoName>{props.repo.name}</RepoName>
-        <SmallText>{props.repo.url}</SmallText>
-      </RepositoryInfoContainer>
-    </RepositoryInfoBackground>
   )
 }
